refactor(e2e): #220 split federated query builders Signed-off-by: Ruoshi Li <iceboundrock@gmail.com>
