+++ internal/e2e_harness/federated/query.go
@@ -8,9 +8,8 @@ import (
 	"strings"
 	"time"
 
-	"github.com/lychee-technology/forma/internal/model"
-
 	"github.com/google/uuid"
+	"github.com/lychee-technology/forma/internal/model"
 )
 
 const (
@@ -275,454 +274,6 @@ func isFederatedTierFileError(err error) bool {
 	return strings.Contains(err.Error(), "No files found") || strings.Contains(err.Error(), "does not exist")
 }
 
-func (h *FederatedTestHarness) buildHotTierQuery(pgConnStr string, schemaID int16, rowIDFilter, attributeFilter, timeWindowFilter string, benchmarkProjection, tradeTimeOnlyProjection bool) string {
-	if benchmarkProjection {
-		if tradeTimeOnlyProjection && schemaID == benchmarkSchemaIDTrade {
-			return h.buildHotTradeTimeOnlyQuery(pgConnStr, schemaID, rowIDFilter)
-		}
-		return h.buildHotTierQueryTargeted(pgConnStr, schemaID, rowIDFilter, attributeFilter, timeWindowFilter)
-	}
-	return fmt.Sprintf(`
-		SELECT 
-			cl.row_id::VARCHAR as row_id,
-			cl.schema_id,
-			cl.changed_at,
-			cl.deleted_at,
-			'' as name,
-			0 as version,
-			'hot' as tier
-		FROM postgres_scan('%s', 'public', 'change_log') cl
-		WHERE cl.flushed_at = 0 
-			AND cl.schema_id = %d
-			%s
-			%s`, pgConnStr, schemaID, rowIDFilter, timeWindowFilter)
-}
-
-func buildHotTierQuery(pgConnStr string, schemaID int16, rowIDFilter, attributeFilter, timeWindowFilter string, benchmarkProjection, tradeTimeOnlyProjection bool) string {
-	return (*FederatedTestHarness)(nil).buildHotTierQuery(pgConnStr, schemaID, rowIDFilter, attributeFilter, timeWindowFilter, benchmarkProjection, tradeTimeOnlyProjection)
-}
-
-func (h *FederatedTestHarness) buildHotTradeTimeOnlyQuery(pgConnStr string, schemaID int16, rowIDFilter string) string {
-	tradeTimeAttrID := h.benchmarkAttributeID(schemaID, "tradeTime")
-	return fmt.Sprintf(`
-		SELECT 
-			cl.row_id::VARCHAR as row_id,
-			cl.schema_id,
-			cl.changed_at,
-			cl.deleted_at,
-			'' as name,
-			0 as version,
-			'' as symbol,
-			'' as exchange,
-			'' as region,
-			0 as tradeType,
-			COALESCE(hot_vals.trade_time, em.bigint_02, 0) as tradeTime,
-			'hot' as tier
-		FROM postgres_scan('%s', 'public', 'change_log') cl
-		LEFT JOIN postgres_scan('%s', 'public', 'entity_main') em
-			ON em.ltbase_schema_id = cl.schema_id AND em.ltbase_row_id::VARCHAR = cl.row_id::VARCHAR
-		LEFT JOIN (
-			SELECT row_id::VARCHAR as row_id, schema_id,
-				MAX(CASE WHEN attr_id = %d THEN value_numeric::BIGINT END) AS trade_time
-			FROM postgres_scan('%s', 'public', 'eav_data')
-			WHERE attr_id = %d
-			GROUP BY schema_id, row_id
-		) hot_vals ON hot_vals.schema_id = cl.schema_id AND hot_vals.row_id = cl.row_id::VARCHAR
-		WHERE cl.flushed_at = 0 
-			AND cl.schema_id = %d
-			%s`, pgConnStr, pgConnStr, tradeTimeAttrID, pgConnStr, tradeTimeAttrID, schemaID, rowIDFilter)
-}
-
-func usesBenchmarkProjectionForSelect(opts *QueryOptions) bool {
-	if opts == nil {
-		return false
-	}
-	if requiresBenchmarkProjectedFilters(opts) {
-		return true
-	}
-	if opts.SortBy != "" && opts.SortBy != "row_id" {
-		return true
-	}
-	return false
-}
-
-func usesBenchmarkProjectionForCount(opts *QueryOptions) bool {
-	if requiresBenchmarkProjectedFilters(opts) {
-		return true
-	}
-	if opts == nil {
-		return false
-	}
-	if opts.Offset <= 0 {
-		return usesBenchmarkProjectionForSelect(opts)
-	}
-	return false
-}
-
-func usesTradeTimeOnlyBenchmarkProjectionForSelect(opts *QueryOptions) bool {
-	if !usesBenchmarkProjectionForSelect(opts) || opts == nil {
-		return false
-	}
-	if opts.SortBy != "tradeTime" || opts.Filter != nil {
-		return false
-	}
-	return opts.TradeTimeStart == 0 && opts.TradeTimeEnd == 0
-}
-
-type hotTierEAVMapping struct {
-	attrIDList   string
-	pivotColumns string
-	selectExprs  string
-	nameExpr     string
-}
-
-func (h *FederatedTestHarness) benchmarkAttributeID(schemaID int16, name string) int {
-	if h != nil && h.Registry != nil {
-		if _, cache, err := h.Registry.GetSchemaAttributeCacheByID(schemaID); err == nil && cache != nil {
-			if metadata, ok := cache[name]; ok {
-				return int(metadata.AttributeID)
-			}
-		}
-	}
-	hash := uint32(2166136261)
-	input := fmt.Sprintf("%d:%s", schemaID, name)
-	for i := 0; i < len(input); i++ {
-		hash ^= uint32(input[i])
-		hash *= 16777619
-	}
-	return int(hash%30000) + 1
-}
-
-func (h *FederatedTestHarness) hotTierEAVMappingForSchema(schemaID int16) hotTierEAVMapping {
-	switch schemaID {
-	case benchmarkSchemaIDTrade:
-		symbolID := h.benchmarkAttributeID(schemaID, "symbol")
-		exchangeID := h.benchmarkAttributeID(schemaID, "exchange")
-		regionID := h.benchmarkAttributeID(schemaID, "region")
-		tradeTypeID := h.benchmarkAttributeID(schemaID, "tradeType")
-		tradeTimeID := h.benchmarkAttributeID(schemaID, "tradeTime")
-		return hotTierEAVMapping{
-			attrIDList: fmt.Sprintf("%d, %d, %d, %d, %d", symbolID, exchangeID, regionID, tradeTypeID, tradeTimeID),
-			pivotColumns: fmt.Sprintf(
-				"MAX(CASE WHEN attr_id = %d THEN value_text END) AS symbol,\n\t\t\t"+
-					"MAX(CASE WHEN attr_id = %d THEN value_text END) AS exchange,\n\t\t\t"+
-					"MAX(CASE WHEN attr_id = %d THEN value_text END) AS region,\n\t\t\t"+
-					"MAX(CASE WHEN attr_id = %d THEN value_numeric::BIGINT END) AS tradeType,\n\t\t\t"+
-					"MAX(CASE WHEN attr_id = %d THEN value_numeric::BIGINT END) AS tradeTime",
-				symbolID, exchangeID, regionID, tradeTypeID, tradeTimeID),
-			selectExprs: "COALESCE(hot_vals.symbol, em.text_01) as symbol,\n\t\t\t" +
-				"COALESCE(hot_vals.exchange, '') as exchange,\n\t\t\t" +
-				"COALESCE(hot_vals.region, em.text_02) as region,\n\t\t\t" +
-				"COALESCE(hot_vals.tradeType, em.smallint_01) as tradeType,\n\t\t\t" +
-				"COALESCE(hot_vals.tradeTime, em.bigint_02) as tradeTime",
-			nameExpr: "COALESCE(hot_vals.symbol, em.text_01, '')",
-		}
-	case benchmarkSchemaIDCustomer:
-		regionID := h.benchmarkAttributeID(schemaID, "region")
-		nameID := h.benchmarkAttributeID(schemaID, "name")
-		return hotTierEAVMapping{
-			attrIDList: fmt.Sprintf("%d, %d", regionID, nameID),
-			pivotColumns: fmt.Sprintf(
-				"MAX(CASE WHEN attr_id = %d THEN value_text END) AS region,\n\t\t\t"+
-					"MAX(CASE WHEN attr_id = %d THEN value_text END) AS name",
-				regionID, nameID),
-			selectExprs: "'' as symbol,\n\t\t\t" +
-				"'' as exchange,\n\t\t\t" +
-				"COALESCE(hot_vals.region, em.text_02) as region,\n\t\t\t" +
-				"0 as tradeType,\n\t\t\t" +
-				"0 as tradeTime",
-			nameExpr: "COALESCE(hot_vals.name, '')",
-		}
-	case benchmarkSchemaIDSecurity:
-		symbolID := h.benchmarkAttributeID(schemaID, "symbol")
-		nameID := h.benchmarkAttributeID(schemaID, "companyName")
-		return hotTierEAVMapping{
-			attrIDList: fmt.Sprintf("%d, %d", symbolID, nameID),
-			pivotColumns: fmt.Sprintf(
-				"MAX(CASE WHEN attr_id = %d THEN value_text END) AS symbol,\n\t\t\t"+
-					"MAX(CASE WHEN attr_id = %d THEN value_text END) AS name",
-				symbolID, nameID),
-			selectExprs: "COALESCE(hot_vals.symbol, em.text_01) as symbol,\n\t\t\t" +
-				"'' as exchange,\n\t\t\t" +
-				"'' as region,\n\t\t\t" +
-				"0 as tradeType,\n\t\t\t" +
-				"0 as tradeTime",
-			nameExpr: "COALESCE(hot_vals.name, hot_vals.symbol, '')",
-		}
-	default:
-		return hotTierEAVMapping{}
-	}
-}
-
-func hotTierEAVMappingForSchema(schemaID int16) hotTierEAVMapping {
-	return (*FederatedTestHarness)(nil).hotTierEAVMappingForSchema(schemaID)
-}
-
-func (h *FederatedTestHarness) buildHotTierQueryTargeted(pgConnStr string, schemaID int16, rowIDFilter, attributeFilter, timeWindowFilter string) string {
-	m := h.hotTierEAVMappingForSchema(schemaID)
-	return fmt.Sprintf(`
-		SELECT 
-			cl.row_id::VARCHAR as row_id,
-			cl.schema_id,
-			cl.changed_at,
-			cl.deleted_at,
-			%s as name,
-			0 as version,
-			%s,
-			'hot' as tier
-		FROM postgres_scan('%s', 'public', 'change_log') cl
-		LEFT JOIN postgres_scan('%s', 'public', 'entity_main') em
-			ON em.ltbase_schema_id = cl.schema_id AND em.ltbase_row_id::VARCHAR = cl.row_id::VARCHAR
-		LEFT JOIN (
-			SELECT row_id::VARCHAR as row_id, schema_id,
-				%s
-			FROM postgres_scan('%s', 'public', 'eav_data')
-			WHERE attr_id IN (%s)
-			GROUP BY schema_id, row_id
-		) hot_vals ON hot_vals.schema_id = cl.schema_id AND hot_vals.row_id = cl.row_id::VARCHAR
-		WHERE cl.flushed_at = 0 
-			AND cl.schema_id = %d
-			%s
-			%s
-			%s`,
-		m.nameExpr, m.selectExprs,
-		pgConnStr, pgConnStr,
-		m.pivotColumns, pgConnStr, m.attrIDList,
-		schemaID, rowIDFilter, attributeFilter, timeWindowFilter)
-}
-
-func buildHotTierQueryTargeted(pgConnStr string, schemaID int16, rowIDFilter, attributeFilter, timeWindowFilter string) string {
-	return (*FederatedTestHarness)(nil).buildHotTierQueryTargeted(pgConnStr, schemaID, rowIDFilter, attributeFilter, timeWindowFilter)
-}
-
-func needsBenchmarkDuckDBMacros(opts *QueryOptions, benchmarkProjection, tradeTimeOnlyProjection bool) bool {
-	return false
-}
-
-func requiresBenchmarkProjectedFilters(opts *QueryOptions) bool {
-	if opts == nil {
-		return false
-	}
-	if opts.TradeTimeStart > 0 || opts.TradeTimeEnd > 0 {
-		return true
-	}
-	if opts.Filter == nil {
-		return false
-	}
-	for key := range opts.Filter.Conditions {
-		if benchmarkQueryColumn(key) != "" {
-			return true
-		}
-	}
-	return false
-}
-
-// buildDirtyExclusion builds the dirty ID exclusion clause.
-func buildDirtyExclusion(dirtyIDs []uuid.UUID) string {
-	if len(dirtyIDs) == 0 {
-		return ""
-	}
-	ids := make([]string, len(dirtyIDs))
-	for i, id := range dirtyIDs {
-		ids[i] = fmt.Sprintf("'%s'", id.String())
-	}
-	return fmt.Sprintf("AND row_id NOT IN (%s)", strings.Join(ids, ","))
-}
-
-// buildRowIDFilter builds the row ID filter clause.
-func buildRowIDFilter(opts *QueryOptions) string {
-	if opts.Filter != nil && opts.Filter.RowID != uuid.Nil {
-		return fmt.Sprintf("AND row_id = '%s'", opts.Filter.RowID.String())
-	}
-	return ""
-}
-
-func buildHotRowIDFilter(opts *QueryOptions) string {
-	if opts.Filter != nil && opts.Filter.RowID != uuid.Nil {
-		return fmt.Sprintf("AND cl.row_id = '%s'", opts.Filter.RowID.String())
-	}
-	return ""
-}
-
-func buildAttributeFilterClause(opts *QueryOptions) string {
-	if opts == nil || opts.Filter == nil || len(opts.Filter.Conditions) == 0 {
-		return ""
-	}
-	parts := make([]string, 0, len(opts.Filter.Conditions))
-	for key, value := range opts.Filter.Conditions {
-		column := benchmarkQueryColumn(key)
-		if column == "" {
-			continue
-		}
-		parts = append(parts, fmt.Sprintf("AND %s = %s", column, benchmarkSQLLiteral(value)))
-	}
-	return strings.Join(parts, " ")
-}
-
-func buildTradeTimeFilterClause(opts *QueryOptions) string {
-	if opts == nil {
-		return ""
-	}
-	parts := make([]string, 0, 2)
-	expression := parquetTradeTimeFilterExpression()
-	if opts.TradeTimeStart > 0 {
-		parts = append(parts, fmt.Sprintf("AND %s >= %d", expression, opts.TradeTimeStart))
-	}
-	if opts.TradeTimeEnd > 0 {
-		parts = append(parts, fmt.Sprintf("AND %s <= %d", expression, opts.TradeTimeEnd))
-	}
-	return strings.Join(parts, " ")
-}
-
-func parquetTradeTimeFilterExpression() string {
-	return "tradeTime"
-}
-
-func buildHotTradeTimeFilterClauseTargeted(opts *QueryOptions) string {
-	if opts == nil {
-		return ""
-	}
-	parts := make([]string, 0, 2)
-	expression := targetedHotFilterExpression("tradeTime")
-	if opts.TradeTimeStart > 0 {
-		parts = append(parts, fmt.Sprintf("AND %s >= %d", expression, opts.TradeTimeStart))
-	}
-	if opts.TradeTimeEnd > 0 {
-		parts = append(parts, fmt.Sprintf("AND %s <= %d", expression, opts.TradeTimeEnd))
-	}
-	return strings.Join(parts, " ")
-}
-
-func benchmarkQueryColumn(attribute string) string {
-	switch attribute {
-	case "symbol":
-		return "symbol"
-	case "exchange":
-		return "exchange"
-	case "region":
-		return "region"
-	case "tradeType":
-		return "tradeType"
-	case "tradeTime":
-		return "tradeTime"
-	default:
-		return ""
-	}
-}
-
-func benchmarkSQLLiteral(value any) string {
-	switch v := value.(type) {
-	case string:
-		return fmt.Sprintf("'%s'", strings.ReplaceAll(v, "'", "''"))
-	case int:
-		return fmt.Sprintf("%d", v)
-	case int64:
-		return fmt.Sprintf("%d", v)
-	case float64:
-		return fmt.Sprintf("%v", v)
-	default:
-		return fmt.Sprintf("'%v'", v)
-	}
-}
-
-func (h *FederatedTestHarness) benchmarkAttrNameSQLCase() string {
-	type attrKey struct {
-		schemaID int16
-		name     string
-	}
-	mapping := map[attrKey]int{
-		{benchmarkSchemaIDTrade, "symbol"}:    h.benchmarkAttributeID(benchmarkSchemaIDTrade, "symbol"),
-		{benchmarkSchemaIDTrade, "exchange"}:  h.benchmarkAttributeID(benchmarkSchemaIDTrade, "exchange"),
-		{benchmarkSchemaIDTrade, "region"}:    h.benchmarkAttributeID(benchmarkSchemaIDTrade, "region"),
-		{benchmarkSchemaIDTrade, "tradeType"}: h.benchmarkAttributeID(benchmarkSchemaIDTrade, "tradeType"),
-		{benchmarkSchemaIDTrade, "tradeTime"}: h.benchmarkAttributeID(benchmarkSchemaIDTrade, "tradeTime"),
-		{benchmarkSchemaIDCustomer, "region"}: h.benchmarkAttributeID(benchmarkSchemaIDCustomer, "region"),
-		{benchmarkSchemaIDCustomer, "name"}:   h.benchmarkAttributeID(benchmarkSchemaIDCustomer, "name"),
-		{benchmarkSchemaIDSecurity, "symbol"}: h.benchmarkAttributeID(benchmarkSchemaIDSecurity, "symbol"),
-		{benchmarkSchemaIDSecurity, "name"}:   h.benchmarkAttributeID(benchmarkSchemaIDSecurity, "companyName"),
-	}
-	parts := make([]string, 0, len(mapping))
-	for key, attrID := range mapping {
-		parts = append(parts, fmt.Sprintf("WHEN schema_id = %d AND attr_id = %d THEN '%s'", key.schemaID, attrID, key.name))
-	}
-	return strings.Join(parts, " ")
-}
-func (h *FederatedTestHarness) benchmarkFunctionsSQL() string {
-	return fmt.Sprintf(`
-		CREATE OR REPLACE MACRO benchmark_attr_name(schema_id, attr_id) AS (
-			CASE %s ELSE '' END
-		);
-		CREATE OR REPLACE MACRO benchmark_text(attr_map, attr_name, fallback_value) AS (
-			COALESCE(CAST(element_at(attr_map, attr_name) AS VARCHAR), fallback_value)
-		);
-		CREATE OR REPLACE MACRO benchmark_name(attr_map) AS (
-			COALESCE(CAST(element_at(attr_map, 'name') AS VARCHAR), CAST(element_at(attr_map, 'symbol') AS VARCHAR), '')
-		);
-		CREATE OR REPLACE MACRO benchmark_int(attr_map, attr_name, fallback_value) AS (
-			COALESCE(TRY_CAST(element_at(attr_map, attr_name) AS INTEGER), fallback_value)
-		);
-		CREATE OR REPLACE MACRO benchmark_bigint(attr_map, attr_name, fallback_value) AS (
-			COALESCE(TRY_CAST(element_at(attr_map, attr_name) AS BIGINT), fallback_value)
-		);
-	`, h.benchmarkAttrNameSQLCase())
-}
-
-func prepareBenchmarkDuckDBMacros(ctx context.Context, h *FederatedTestHarness) error {
-	_, err := h.Duck.DB.ExecContext(ctx, h.benchmarkFunctionsSQL())
-	if err != nil {
-		return fmt.Errorf("prepare benchmark duckdb macros: %w", err)
-	}
-	return nil
-}
-
-func buildHotAttributeFilterClauseTargeted(opts *QueryOptions) string {
-	if opts == nil || opts.Filter == nil || len(opts.Filter.Conditions) == 0 {
-		return ""
-	}
-	parts := make([]string, 0, len(opts.Filter.Conditions))
-	for key, value := range opts.Filter.Conditions {
-		expression := targetedHotFilterExpression(key)
-		if expression == "" {
-			continue
-		}
-		parts = append(parts, fmt.Sprintf("AND %s = %s", expression, benchmarkSQLLiteral(value)))
-	}
-	return strings.Join(parts, " ")
-}
-
-func targetedHotFilterExpression(attribute string) string {
-	switch attribute {
-	case "symbol":
-		return "COALESCE(hot_vals.symbol, em.text_01)"
-	case "exchange":
-		return "COALESCE(hot_vals.exchange, '')"
-	case "region":
-		return "COALESCE(hot_vals.region, em.text_02)"
-	case "tradeType":
-		return "COALESCE(hot_vals.tradeType, em.smallint_01)"
-	case "tradeTime":
-		return "COALESCE(hot_vals.tradeTime, em.bigint_02)"
-	default:
-		return ""
-	}
-}
-
-func buildOrderByClause(opts *QueryOptions) string {
-	prefix := ""
-	if opts == nil {
-		return prefix + "row_id ASC"
-	}
-	column := benchmarkQueryColumn(opts.SortBy)
-	if column == "" {
-		column = "row_id"
-	}
-	direction := "ASC"
-	if opts.SortDesc {
-		direction = "DESC"
-	}
-	return fmt.Sprintf("%s%s %s, %srow_id ASC", prefix, column, direction, prefix)
-}
-
 // buildPGConnString builds the Postgres connection string for DuckDB.
 func (h *FederatedTestHarness) buildPGConnString() string {
 	host := h.PGHost
@@ -853,202 +404,6 @@ func (h *FederatedTestHarness) ExecutePostgresQuery(ctx context.Context, opts *Q
 		Duration:     time.Since(start),
 		Plan:         buildPostgresOnlyExecutionPlan(time.Since(start), opts.PreferHot),
 	}, nil
-}
-
-func (h *FederatedTestHarness) buildPostgresOnlySelectQuery(opts *QueryOptions) (string, []any) {
-	args := []any{h.SchemaID}
-	attrIDs := h.benchmarkPostgresAttributeIDs()
-	query := strings.Builder{}
-	query.WriteString(`
-		SELECT
-			cl.row_id::VARCHAR,
-			cl.schema_id,
-			cl.changed_at,
-			COALESCE(cl.deleted_at, 0),
-			COALESCE(hot_vals.name, hot_vals.symbol, '') as name,
-			0 as version`)
-	if usesBenchmarkProjectionForSelect(opts) {
-		query.WriteString(`,
-			COALESCE(hot_vals.symbol, em.text_01, '') as symbol,
-			COALESCE(hot_vals.exchange, '') as exchange,
-			COALESCE(hot_vals.region, em.text_02, '') as region,
-			COALESCE(hot_vals.trade_type, em.smallint_01::BIGINT, 0) as tradeType,
-			COALESCE(hot_vals.trade_time, em.bigint_02, 0) as tradeTime`)
-	}
-	query.WriteString(fmt.Sprintf(`
-		FROM change_log cl
-		LEFT JOIN entity_main em
-			ON em.ltbase_schema_id = cl.schema_id AND em.ltbase_row_id = cl.row_id
-		LEFT JOIN (
-			SELECT schema_id, row_id,
-				MAX(CASE WHEN attr_id = %d THEN value_text END) AS symbol,
-				MAX(CASE WHEN attr_id = %d THEN value_text END) AS exchange,
-				MAX(CASE WHEN attr_id = %d THEN value_text END) AS region,
-				MAX(CASE WHEN attr_id = %d THEN value_numeric::BIGINT END) AS trade_type,
-				MAX(CASE WHEN attr_id = %d THEN value_numeric::BIGINT END) AS trade_time,
-				MAX(CASE WHEN attr_id = %d THEN value_text END) AS name
-			FROM eav_data
-			WHERE attr_id IN (%d, %d, %d, %d, %d, %d)
-			GROUP BY schema_id, row_id
-		) hot_vals ON hot_vals.schema_id = cl.schema_id AND hot_vals.row_id = cl.row_id
-		WHERE cl.schema_id = $1 AND cl.flushed_at = 0 AND (cl.deleted_at IS NULL OR cl.deleted_at = 0)`,
-		attrIDs.symbol, attrIDs.exchange, attrIDs.region, attrIDs.tradeType, attrIDs.tradeTime, attrIDs.name,
-		attrIDs.symbol, attrIDs.exchange, attrIDs.region, attrIDs.tradeType, attrIDs.tradeTime, attrIDs.name))
-	filterSQL, filterArgs := buildPostgresOnlyFilterClauses(opts, 2)
-	query.WriteString(filterSQL)
-	args = append(args, filterArgs...)
-	query.WriteString(fmt.Sprintf(" ORDER BY %s LIMIT $%d OFFSET $%d", buildPostgresOnlyOrderBy(opts), len(args)+1, len(args)+2))
-	args = append(args, opts.Limit, opts.Offset)
-	return query.String(), args
-}
-
-func (h *FederatedTestHarness) buildPostgresOnlyCountQuery(opts *QueryOptions) (string, []any) {
-	args := []any{h.SchemaID}
-	attrIDs := h.benchmarkPostgresAttributeIDs()
-	query := strings.Builder{}
-	query.WriteString(fmt.Sprintf(`
-		SELECT COUNT(*)
-		FROM change_log cl
-		LEFT JOIN entity_main em
-			ON em.ltbase_schema_id = cl.schema_id AND em.ltbase_row_id = cl.row_id
-		LEFT JOIN (
-			SELECT schema_id, row_id,
-				MAX(CASE WHEN attr_id = %d THEN value_text END) AS symbol,
-				MAX(CASE WHEN attr_id = %d THEN value_text END) AS exchange,
-				MAX(CASE WHEN attr_id = %d THEN value_text END) AS region,
-				MAX(CASE WHEN attr_id = %d THEN value_numeric::BIGINT END) AS trade_type,
-				MAX(CASE WHEN attr_id = %d THEN value_numeric::BIGINT END) AS trade_time,
-				MAX(CASE WHEN attr_id = %d THEN value_text END) AS name
-			FROM eav_data
-			WHERE attr_id IN (%d, %d, %d, %d, %d, %d)
-			GROUP BY schema_id, row_id
-		) hot_vals ON hot_vals.schema_id = cl.schema_id AND hot_vals.row_id = cl.row_id
-		WHERE cl.schema_id = $1 AND cl.flushed_at = 0 AND (cl.deleted_at IS NULL OR cl.deleted_at = 0)`,
-		attrIDs.symbol, attrIDs.exchange, attrIDs.region, attrIDs.tradeType, attrIDs.tradeTime, attrIDs.name,
-		attrIDs.symbol, attrIDs.exchange, attrIDs.region, attrIDs.tradeType, attrIDs.tradeTime, attrIDs.name))
-	filterSQL, filterArgs := buildPostgresOnlyFilterClauses(opts, 2)
-	query.WriteString(filterSQL)
-	args = append(args, filterArgs...)
-	return query.String(), args
-}
-
-func buildPostgresOnlyFilterClauses(opts *QueryOptions, placeholderStart int) (string, []any) {
-	if opts == nil {
-		return "", nil
-	}
-	args := make([]any, 0)
-	parts := make([]string, 0)
-	placeholder := placeholderStart
-	if opts.Filter != nil && opts.Filter.RowID != uuid.Nil {
-		parts = append(parts, fmt.Sprintf("AND cl.row_id = $%d", placeholder))
-		args = append(args, opts.Filter.RowID)
-		placeholder++
-	}
-	if opts.Filter != nil {
-		for key, value := range opts.Filter.Conditions {
-			expression := postgresOnlyFilterExpression(key)
-			if expression == "" {
-				continue
-			}
-			parts = append(parts, fmt.Sprintf("AND %s = $%d", expression, placeholder))
-			args = append(args, value)
-			placeholder++
-		}
-	}
-	if opts.TradeTimeStart > 0 {
-		parts = append(parts, fmt.Sprintf("AND %s >= $%d", postgresOnlyFilterExpression("tradeTime"), placeholder))
-		args = append(args, opts.TradeTimeStart)
-		placeholder++
-	}
-	if opts.TradeTimeEnd > 0 {
-		parts = append(parts, fmt.Sprintf("AND %s <= $%d", postgresOnlyFilterExpression("tradeTime"), placeholder))
-		args = append(args, opts.TradeTimeEnd)
-	}
-	if len(parts) == 0 {
-		return "", nil
-	}
-	return " " + strings.Join(parts, " "), args
-}
-
-func buildPostgresOnlyOrderBy(opts *QueryOptions) string {
-	if opts == nil || opts.SortBy == "" {
-		return "cl.row_id ASC"
-	}
-	column := postgresOnlyFilterExpression(opts.SortBy)
-	if column == "" {
-		column = "cl.row_id"
-	}
-	direction := "ASC"
-	if opts.SortDesc {
-		direction = "DESC"
-	}
-	return fmt.Sprintf("%s %s, cl.row_id ASC", column, direction)
-}
-
-type postgresBenchmarkAttributeIDs struct {
-	symbol    int
-	exchange  int
-	region    int
-	tradeType int
-	tradeTime int
-	name      int
-}
-
-func (h *FederatedTestHarness) benchmarkPostgresAttributeIDs() postgresBenchmarkAttributeIDs {
-	if h != nil && h.Registry != nil {
-		if _, cache, err := h.Registry.GetSchemaAttributeCacheByID(benchmarkSchemaIDTrade); err == nil && cache != nil {
-			ids := postgresBenchmarkAttributeIDs{}
-			if meta, ok := cache["symbol"]; ok {
-				ids.symbol = int(meta.AttributeID)
-			}
-			if meta, ok := cache["exchange"]; ok {
-				ids.exchange = int(meta.AttributeID)
-			}
-			if meta, ok := cache["region"]; ok {
-				ids.region = int(meta.AttributeID)
-			}
-			if meta, ok := cache["tradeType"]; ok {
-				ids.tradeType = int(meta.AttributeID)
-			}
-			if meta, ok := cache["tradeTime"]; ok {
-				ids.tradeTime = int(meta.AttributeID)
-			}
-			return ids
-		}
-	}
-	return postgresBenchmarkAttributeIDs{
-		symbol:    h.benchmarkAttributeID(benchmarkSchemaIDTrade, "symbol"),
-		exchange:  h.benchmarkAttributeID(benchmarkSchemaIDTrade, "exchange"),
-		region:    h.benchmarkAttributeID(benchmarkSchemaIDTrade, "region"),
-		tradeType: h.benchmarkAttributeID(benchmarkSchemaIDTrade, "tradeType"),
-		tradeTime: h.benchmarkAttributeID(benchmarkSchemaIDTrade, "tradeTime"),
-		name:      h.benchmarkAttributeID(benchmarkSchemaIDTrade, "name"),
-	}
-}
-
-func postgresOnlyFilterExpression(attribute string) string {
-	switch attribute {
-	case "symbol":
-		return "COALESCE(hot_vals.symbol, em.text_01, '')"
-	case "exchange":
-		return "COALESCE(hot_vals.exchange, '')"
-	case "region":
-		return "COALESCE(hot_vals.region, em.text_02, '')"
-	case "tradeType":
-		return "COALESCE(hot_vals.trade_type, em.smallint_01::BIGINT, 0)"
-	case "tradeTime":
-		return "COALESCE(hot_vals.trade_time, em.bigint_02, 0)"
-	default:
-		return ""
-	}
-}
-
-func buildPostgresOnlyExecutionPlan(duration time.Duration, preferHot bool) *model.ExecutionPlan {
-	notes := []string{"hot_buffer_scanned", "postgres_only_execution"}
-	if preferHot {
-		notes = append(notes, "prefer_hot_override")
-	}
-	return &model.ExecutionPlan{Notes: notes, Timings: map[string]int64{"total": duration.Milliseconds()}}
 }
 
 // StreamFederatedQuery streams query results with a handler callback.
